+++ quickcheck/pom.xml
@@ -6,7 +6,6 @@
     <version>1.0</version>
     <packaging>pom</packaging>
     <modules>
-        <module>quickcheckmodel</module>
         <module>quickcheckweb</module>
     </modules>
     <properties>
