--- quickcheck/quickcheckmodel/src/main/java/dto/MedicoDTO.java
@@ -1,118 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package dto;
-
-import java.sql.Date;
-
-/**
- *
- * @author Usuário
- */
-public class MedicoDTO {
-    private String nome, senha, endereco, email, crm, cpf, telefone;
-    private Date dataNascimento;
-    private Integer codigo;
-    
-    public MedicoDTO() {
-    }
-
-public MedicoDTO(String nome, String senha, String endereco, String email, String convenio, String cpf, String telefone,
-            Date datanascimento, Integer codigo) {
-        this.nome = nome;
-        this.senha = senha;
-        this.endereco = endereco;
-        this.email = email;
-        this.crm = crm;
-        this.cpf = cpf;
-        this.telefone = telefone;
-        this.dataNascimento = dataNascimento;
-        this.codigo = codigo;
-    }
-
-    public MedicoDTO(String nome, String senha, String endereco, String email, String convenio, String cpf, String telefone, Date datanascimento) {
-        this.nome = nome;
-        this.senha = senha;
-        this.endereco = endereco;
-        this.email = email;
-        this.crm = crm;
-        this.cpf = cpf;
-        this.telefone = telefone;
-        this.dataNascimento = dataNascimento;
-    }
-    
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-    
-    public String getSenha() {
-        return senha;
-    }
-
-    public void setSenha(String senha) {
-        this.senha = senha;
-    }
-
-    public String getEndereco() {
-        return endereco;
-    }
-
-    public void setEndereco(String endereco) {
-        this.endereco = endereco;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getCrm() {
-        return crm;
-    }
-
-    public void setCrm(String convenio) {
-        this.crm = convenio;
-    }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-
-    public String getTelefone() {
-        return telefone;
-    }
-
-    public void setTelefone(String telefone) {
-        this.telefone = telefone;
-    }
-
-    public Date getDataNascimento() {
-        return dataNascimento;
-    }
-
-    public void setDataNascimento(Date datanascimento) {
-        this.dataNascimento = datanascimento;
-    }
-
-    public Integer getCodigo() {
-        return codigo;
-    }
-
-    public void setCodigo(Integer codigo) {
-        this.codigo = codigo;
-    }
-
-}
-
