--- quickcheck/quickcheckmodel/src/main/java/dto/PacienteDTO.java
@@ -1,100 +1,0 @@
-package dto;
-
-import java.sql.Date;
-
-public class PacienteDTO {
-    private String nome, endereco, email, convenio, cpf, telefone;
-    private Date datanascimento;
-    private Integer codigo;
-
-    public PacienteDTO() {
-    }
-   
-    public PacienteDTO(String nome, String endereco, String email, String convenio, String cpf, String telefone,
-            Date datanascimento, Integer codigo) {
-        this.nome = nome;
-        this.endereco = endereco;
-        this.email = email;
-        this.convenio = convenio;
-        this.cpf = cpf;
-        this.telefone = telefone;
-        this.datanascimento = datanascimento;
-        this.codigo = codigo;
-    }
-
-    public PacienteDTO(String nome, String endereco, String email, String convenio, String cpf, String telefone, Date datanascimento) {
-        this.nome = nome;
-        this.endereco = endereco;
-        this.email = email;
-        this.convenio = convenio;
-        this.cpf = cpf;
-        this.telefone = telefone;
-        this.datanascimento = datanascimento;
-    }
-
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getEndereco() {
-        return endereco;
-    }
-
-    public void setEndereco(String endereco) {
-        this.endereco = endereco;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getConvenio() {
-        return convenio;
-    }
-
-    public void setConvenio(String convenio) {
-        this.convenio = convenio;
-    }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-
-    public String getTelefone() {
-        return telefone;
-    }
-
-    public void setTelefone(String telefone) {
-        this.telefone = telefone;
-    }
-
-    public Date getDatanascimento() {
-        return datanascimento;
-    }
-
-    public void setDatanascimento(Date datanascimento) {
-        this.datanascimento = datanascimento;
-    }
-
-    public Integer getCodigo() {
-        return codigo;
-    }
-
-    public void setCodigo(Integer codigo) {
-        this.codigo = codigo;
-    }
-
-}
