+++ quickcheck/quickcheckweb/pom.xml
@@ -34,13 +34,13 @@
             <version>20191001.1</version>
         </dependency>
         <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>quickcheck-model</artifactId>
-            <version>${project.version}</version>
+            <groupId>mysql</groupId>
+            <artifactId>mysql-connector-java</artifactId>
+            <version>8.0.28</version>
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>quickcheckmodel</artifactId>
+            <artifactId>quickcheck-model</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
@@ -52,6 +52,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.10.1</version>
                 <configuration>
+                    <source>11</source>
+                    <target>11</target>
                     <compilerArgs>
                         <args>-Djava.endorsed.dirs=${endorsed.dir}</args>
                     </compilerArgs>
