+++ quickcheck/quickcheckweb/src/main/java/model/dao/MedicoDAO.java
@@ -2,10 +2,11 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package dao;
+package model.dao;
 
-import db.DBConnector;
-import dto.MedicoDTO;
+import model.db.DBConnector;
+import model.dto.MedicoDTO;
+
 import java.math.BigInteger;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
