+++ quickcheck/quickcheckweb/src/main/java/model/db/DBConnector.java
@@ -1,4 +1,4 @@
-package db;
+package model.db;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -9,22 +9,29 @@ public class DBConnector {
 
     private DBConnector(){}
 
-    // Strings com informações para conexão ao BD
-    private static final String URL = "jdbc:mysql://us-cdbr-east-06.cleardb.net:3306/heroku_aa28e844ee07386?autoReconnect=true&useSSL=false";
-    private static final String USUARIO = "b918cc3160b707";
-    private static final String SENHA = "16d9b4b7";
-    
-    // Tenta estabelecer conexão com o banco de dados.
+
     public static Connection getConexao() throws SQLException, ClassNotFoundException {
         try {
            Class.forName("com.mysql.cj.jdbc.Driver");
-            return DriverManager.getConnection(URL, USUARIO, SENHA);
+            return DriverManager.getConnection("jdbc:mysql://database-1.cty2gsaowhzj.us-east-2.rds.amazonaws.com:3306/quickcheck?zeroDateTimeBehavior=CONVERT_TO_NULL&useSSL=false", "admin", "12345678");
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
 	}
-    
-    // Fecha conexão com o banco de dados imediatamente, ao invés de esperar o conector realizar a tarefa.
+
+    public static boolean testarConexao() {
+        try {
+            Connection conexao = getConexao();
+            if (conexao != null) {
+                conexao.close();
+                return true;
+            }
+        } catch (SQLException | ClassNotFoundException e) {
+            System.out.println("Erro ao testar a conexão: " + e.getMessage());
+        }
+        return false;
+    }
+        
     public static void fecharConexao (Connection conexao, Statement comando) throws SQLException{
         if(comando != null) {
             comando.close();
