+++ quickcheck/quickcheckweb/src/main/java/model/service/MedicoService.java
@@ -1,0 +1,24 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package model.service;
+
+
+import java.security.NoSuchAlgorithmException;
+import java.sql.Date;
+import java.sql.SQLException;
+
+import model.dao.MedicoDAO;
+import model.dto.MedicoDTO;
+
+public class MedicoService {
+    public static int cadastrarMedico(String nome, String senha, String endereco, String email, String crm, String cpf, String telefone,
+    Date datanascimento) throws NoSuchAlgorithmException, SQLException, ClassNotFoundException {
+        
+       MedicoDTO medicoDTO = new MedicoDTO (nome, senha, endereco, email, crm, cpf, telefone, datanascimento);
+      
+       return MedicoDAO.inserirMedico(medicoDTO);
+        
+    }
+}
