+++ quickcheck/quickcheckweb/src/main/java/sistemadiagnostico/quickcheck/web/PacienteBean.java
@@ -5,16 +5,26 @@ import jakarta.faces.bean.SessionScoped;
 import java.util.ArrayList;
 import java.util.List;
 
+import model.db.DBConnector;
+import model.dto.PacienteDTO;
+import model.service.PacienteService;
+
 @SessionScoped
 @ManagedBean
 public class PacienteBean {
     private PacienteDTO paciente = new PacienteDTO();
+    private PacienteService pacienteService = new PacienteService();
+
     private List<PacienteDTO> pacientes = new ArrayList<>();
 
     public void inserirPaciente() {
-        System.out.println(paciente.getEmail());
-       pacientes.add(paciente);
-       paciente = new PacienteDTO();
+        pacienteService.cadastrarPaciente(paciente);
+        pacientes = pacienteService.listar();
+        paciente = new PacienteDTO();
+    }
+
+    public void listar() {
+        pacientes = pacienteService.listar();
     }
 
     public PacienteDTO getPaciente() {
