+++ quickcheck/quickcheckweb/src/main/webapp/WEB-INF/glassfish-resources.xml
@@ -2,13 +2,23 @@
 <!DOCTYPE resources PUBLIC "-//GlassFish.org//DTD GlassFish Application Server 3.1 Resource Definitions//EN" "http://glassfish.org/dtds/glassfish-resources_1_5.dtd">
 <resources>
     <jdbc-connection-pool allow-non-component-callers="false" associate-with-thread="false" connection-creation-retry-attempts="0" connection-creation-retry-interval-in-seconds="10" connection-leak-reclaim="false" connection-leak-timeout-in-seconds="0" connection-validation-method="auto-commit" datasource-classname="com.mysql.jdbc.jdbc2.optional.MysqlDataSource" fail-all-connections="false" idle-timeout-in-seconds="300" is-connection-validation-required="false" is-isolation-level-guaranteed="true" lazy-connection-association="false" lazy-connection-enlistment="false" match-connections="false" max-connection-usage-count="0" max-pool-size="32" max-wait-time-in-millis="60000" name="mysql_heroku_aa28e844ee07386_b918cc3160b707Pool" non-transactional-connections="false" pool-resize-quantity="2" res-type="javax.sql.DataSource" statement-timeout-in-seconds="-1" steady-pool-size="8" validate-atmost-once-period-in-seconds="0" wrap-jdbc-objects="false">
-        <property name="serverName" value="us-cdbr-east-06.cleardb.net"/>
+        <property name="serverName" value="database-1.cty2gsaowhzj.us-east-2.rds.amazonaws.com"/>
         <property name="portNumber" value="3306"/>
-        <property name="databaseName" value="heroku_aa28e844ee07386"/>
-        <property name="User" value="b918cc3160b707"/>
-        <property name="Password" value="16d9b4b7"/>
+        <property name="databaseName" value="quickcheck"/>
+        <property name="User" value="admin"/>
+        <property name="Password" value="12345678"/>
         <property name="URL" value="jdbc:mysql://us-cdbr-east-06.cleardb.net:3306/heroku_aa28e844ee07386?autoReconnect=true&amp;useSSL=false"/>
         <property name="driverClass" value="com.mysql.cj.jdbc.Driver"/>
     </jdbc-connection-pool>
     <jdbc-resource enabled="true" jndi-name="java:app/heroku" object-type="user" pool-name="mysql_heroku_aa28e844ee07386_b918cc3160b707Pool"/>
+    <jdbc-connection-pool allow-non-component-callers="false" associate-with-thread="false" connection-creation-retry-attempts="0" connection-creation-retry-interval-in-seconds="10" connection-leak-reclaim="false" connection-leak-timeout-in-seconds="0" connection-validation-method="auto-commit" datasource-classname="com.mysql.jdbc.jdbc2.optional.MysqlDataSource" fail-all-connections="false" idle-timeout-in-seconds="300" is-connection-validation-required="false" is-isolation-level-guaranteed="true" lazy-connection-association="false" lazy-connection-enlistment="false" match-connections="false" max-connection-usage-count="0" max-pool-size="32" max-wait-time-in-millis="60000" name="mysql_quickcheck_adminPool" non-transactional-connections="false" pool-resize-quantity="2" res-type="javax.sql.DataSource" statement-timeout-in-seconds="-1" steady-pool-size="8" validate-atmost-once-period-in-seconds="0" wrap-jdbc-objects="false">
+        <property name="serverName" value="database-1.cty2gsaowhzj.us-east-2.rds.amazonaws.com"/>
+        <property name="portNumber" value="3306"/>
+        <property name="databaseName" value="quickcheck"/>
+        <property name="User" value="admin"/>
+        <property name="Password" value="12345678"/>
+        <property name="URL" value="jdbc:mysql://database-1.cty2gsaowhzj.us-east-2.rds.amazonaws.com:3306/quickcheck?zeroDateTimeBehavior=CONVERT_TO_NULL"/>
+        <property name="driverClass" value="com.mysql.cj.jdbc.Driver"/>
+    </jdbc-connection-pool>
+    <jdbc-resource enabled="true" jndi-name="java:app/AWS" object-type="user" pool-name="mysql_quickcheck_adminPool"/>
 </resources>
