+++ quickcheck/quickcheckweb/src/main/webapp/login.xhtml
@@ -1,14 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE html>
-<html>
+<html xmlns="http://www.w3.org/1999/xhtml" xmlns:h="http://xmlns.jcp.org/jsf/html"
+    xmlns:f="http://xmlns.jcp.org/jsf/core" xmlns:jsf="http://xmlns.jcp.org/jsf"
+    xmlns:ui="http://xmlns.jcp.org/jsf/facelets" xmlns:p="http://primefaces.org/ui">
+
 <head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <meta charset="UTF-8" />
+
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+
     <title>Document</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous" />
 
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ENjdO4Dr2bkBIFxQpeoTz1HIcje39Wm4jDKdf19U8gI4ddQ3GYNS7NTKfAdVQSZe" crossorigin="anonymous"></script>
-
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-ENjdO4Dr2bkBIFxQpeoTz1HIcje39Wm4jDKdf19U8gI4ddQ3GYNS7NTKfAdVQSZe"
+        crossorigin="anonymous"></script>
     <style>
         * {
             margin: 0;
@@ -69,35 +79,44 @@
             font-size: 15px;
         }
     </style>
-
 </head>
-<body>
-    <div class="login">
-        <h1 class="text-center">Seja bem-vindo(a)!</h1>
 
+<body>
+
+    <div class="login">
+
+        <h1 class="text-center">Seja bem-vindo(a)!</h1>
         <form class="needs-validation">
+
             <div class="form-group was-validated">
                 <label class="form-label" for="email">Endereço de email</label>
-                <input class="form-control" type="email" id="email" required>
+                <input class="form-control" type="email" id="email" required="" />
+
                 <div class="invalid-feedback">
                     Favor inserir seu endereço de email.
                 </div>
             </div>
+
             <div class="form-group was-validated">
                 <label class="form-label" for="password">Senha</label>
-                <input class="form-control" type="password" id="password" required>
+                <input class="form-control" type="password" id="password" required="" />
+
                 <div class="invalid-feedback">
                     Favor inserir sua senha.
                 </div>
+
                 <p class="senha"><u>Esqueci minha senha</u></p>
             </div>
+
             <div class="form-group form-check">
-                <input class="form-check-input" type="checkbox" id="check">
+                <input class="form-check-input" type="checkbox" id="check" />
                 <label class="form-check-label" for="check">Lembrar de mim</label>
             </div>
-            <input class="btn btn-success w-100" type="submit" value="Entrar">
+            <input class="btn btn-success w-100" type="submit" value="Entrar" />
+
             <p class="criar">Não tem cadastro? <b><u>Crie um</u></b></p>
         </form>
     </div>
 </body>
+
 </html>
